defensive check if mount exists

+++ rsync-home-dir-to-rabbit.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ ! -d "/Volumes/rabbit" ]; then
+    echo "rabbit is not connected. Exiting."
+    exit 1
+fi
+
 rsync -avh --delete -E \
   --exclude=".DS_Store" \
   --exclude="Library/" \
